Fixes to Coco Data Quality Sample

+++ open-metadata-implementation/adapters/open-connectors/data-store-connectors/file-connectors/csv-file-connector/src/main/java/org/odpi/openmetadata/adapters/connectors/datastore/csvfile/CSVFileStore.java
@@ -50,5 +50,5 @@ public interface CSVFileStore extends BasicFileStore
      * @throws FileException there is a problem accessing the file
      * @throws FileReadException unable to find, open or read the file, or the file does not include the requested record.
      */
-    List<String>      readRecord(int  rowNumber) throws FileException, FileReadException;
+    List<String>      readRecord(long  rowNumber) throws FileException, FileReadException;
 }
